Make TFC buckets all non-discrete fluid containers, able to carry increments of a single bucket. Should solve issues with remainders in barrels. Remove old twig -> torch recipe (Fixes #2170). Fix placed items coercing extracted slots to size = 1, relevant for recipes that produce multiple items (Fixes #2175). Fix aqueducts not being able to carry spring water and add test for expected waters. Tweak thirst decay, reduce it a tad to compensate for no saturation-like stat.
