--- src/main/resources/data/tfc/recipes/heating/torch_from_twig.json
@@ -1,11 +1,0 @@
-{
-  "__comment__": "This file was automatically created by mcresources",
-  "type": "tfc:heating",
-  "ingredient": {
-    "tag": "tfc:twigs"
-  },
-  "result_item": {
-    "item": "tfc:torch"
-  },
-  "temperature": 60
-}
